backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.1
django-shell-plus==1.1.7
djangorestframework==3.14.0
djangorestframework-filters==0.11.1
djangorestframework-simplejwt==5.2.2
isbnlib==3.10.12
isbntools==4.3.28
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
